MC stuck questions added + fixed answer generation
Changes:
If user has 3 or more incorrect answers for question tm will be stored and added the next time MC is loaded. Only used in standard mode.

Answer generation will ensure that if the answer is positive then the other answers will be posibive and if the answer is negative then the other answers will be negative.

Fixed question generation by having it only generate 5 initial questions onCreate() then the rest in a async thread (us 2 threads in total).